server/pyproject.toml
[tool.poetry]
authors = ["admin <admin@example.com>"]
description = "FastAPI project generated using minimal-fastapi-postgres-template."
name = "app"
version = "0.1.0-alpha"

[tool.poetry.dependencies]
python = "^3.12"

alembic = "^1.13.1"
asyncpg = "^0.29.0"
bcrypt = "^4.1.3"
fastapi = "^0.111.0"
pydantic = {extras = ["dotenv", "email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.30"
retell-sdk = "^3.15.1"
fastapi-cors = "^0.0.6"
firebase-admin = "^6.5.0"
stripe = "^9.10.0"
twilio = "^9.2.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.5.1"
freezegun = "^1.5.0"
gevent = "^24.2.1"
httpx = "^0.27.0"
mypy = "^1.10.0"
pre-commit = "^3.7.0"
pytest = "^8.2.0"
# do not bump pytest-asyncio until https://github.com/pytest-dev/pytest-asyncio/issues/706 resolved
pytest-asyncio = "0.21.2"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.6.1"
ruff = "^0.4.3"
types-passlib = "^1.7.7.20240327"
uvicorn = {extras = ["standard"], version = "^0.29.0"}

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "-vv -n auto --cov --cov-report xml --cov-report term-missing --cov-fail-under=100"
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.coverage.run]
concurrency = ["gevent"]
omit = ["app/tests/*"]
source = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
# pycodestyle, pyflakes, isort, pylint, pyupgrade
ignore = ["E501"]
select = ["E", "F", "I", "PL", "UP", "W"]
